refined bootstrap logic, improved install/uninstall scripts, with minor updates to the library scripts

+++ utils/uninstall.sh
@@ -4,16 +4,30 @@
 # A shell script to automate the uninstall of the dotfiles.
 #
 
-# The parent directory of this script
-DOTFILES="${0:a:h}"
-INIT="$DOTFILES/lib/systemd/init"
-[[ -r "$INIT" ]] && source "$INIT" || exit 1
+# Source the initialization script to set up the environment.
+INIT_SCRIPT="${0:a:h}/lib/systemd/init"
+if [[ -r "$INIT_SCRIPT" ]]; then
+  source "${INIT_SCRIPT}"
+else
+  echo "Error: Initialization script ${INIT_SCRIPT} not found to set up the environment." >&2
+  exit 1
+fi
+
+# Confirmation prompt
+read -q "REPLY?This will uninstall the dotfiles and remove related configurations. Are you sure you want to proceed? [y/N] "
+echo ""
+if [[ ! "$REPLY" =~ ^[Yy]$ ]]; then
+  echo "Uninstallation aborted by the user."
+  exit 0
+fi
 
 #
 # Uninstall
 #
 
-# Remove configurations
+echo "Start removing Dotfiles..."
+
+# Remove configurations.
 for file in "$HOME/editorconfig" "$HOME/.{hushlogin,zshenv}"; do
   [[ -f "$file" ]] && rm "$file"
 done
@@ -23,7 +37,7 @@ for dir in "$XDG_CONFIG_HOME/{curl,git,nano,node,npm,ssh,tmux,zsh}" "$XDG_DATA_H
   echo "configurations symlinks removed."
 done
 
-# Remove editorconfig symlink
+# Remove editorconfig symlink.
 if [ -L "${HOME}/.editorconfig" ]; then
   rm "${HOME}/.editorconfig"
   echo "editorconfig symlink removed."
@@ -31,25 +45,30 @@ else
   echo "No editorconfig symlink found in the home directory."
 fi
 
-# Uninstall Launch Agents and Unload .plist files
+# Uninstall Launch Agents and Unload .plist files.
 LAUNCHD_DIR="${LIB_DIR}/launchd"
 LAUNCH_AGENTS="${HOME}/Library/LaunchAgents"
 LAUNCH_DAEMONS="/Library/LaunchDaemons"
 
 if [[ -d "$LAUNCHD_DIR" ]]; then
+  if [[ -d "$LAUNCHD_DIR/norflow" ]]; then
+    launchctl unload "${LAUNCH_AGENTS}/com.shell.Norflow.plist"
+    rm "${LAUNCH_AGENTS}/com.shell.Norflow.plist"
+  fi
+
+  if [[ -d "$LAUNCHD_DIR/prune" ]]; then
+    launchctl unload "${LAUNCH_AGENTS}/com.shell.Prune.plist"
+    rm "${LAUNCH_AGENTS}/com.shell.Prune.plist"
+  fi
+
   if [[ -d "$LAUNCHD_DIR/updates" ]]; then
     launchctl unload "${LAUNCH_AGENTS}/com.shell.Updates.plist"
     rm "${LAUNCH_AGENTS}/com.shell.Updates.plist"
   fi
-
-  if [[ -d "$LAUNCHD_DIR/launchpad" ]]; then
-    sudo launchctl unload -w "${LAUNCH_DAEMONS}/com.shell.Launchpad.plist"
-    sudo rm "${LAUNCH_DAEMONS}/com.shell.Launchpad.plist"
-  fi
 fi
 
-# Unload and remove VSCode configs
-CODE=/Applications/Visual\ Studio\ Code.app
+# Unload and remove VSCode configs.
+CODE="/Applications/Visual\ Studio\ Code.app"
 CODE_USER="$HOME/Library/Application Support/Code/User"
 if [[ -d "$CODE" ]]; then
   for config in "$CODE_USER"/{keybindings.json,settings.json}; do
@@ -61,4 +80,4 @@ if [[ -d "$CODE" ]]; then
   fi
 fi
 
-echo 'Dotfiles uninstalled!'
+echo "Dotfiles uninstalled successfully!"
